+++ envs/dev/views/news/default/detail.php
@@ -1,4 +1,15 @@
-<h1>News Detail</h1>
-<pre>
-<?print_r($model); ?>
-</pre>
+<?php
+/* @var $this \luya\web\View */
+/* @var $model \luya\news\models\Article */
+?>
+<h1><?= $model->title; ?></h1>
+<ol>
+<? foreach ($model->tags as $tag): ?>
+<li><? print_r($tag->toArray()); ?><?= $tag->name; ?></li>
+<? endforeach; ?>
+</ol>
+
+
+<? foreach ($model->findTags() as $tag): ?>
+<?= $tag->name; ?><br />
+<? endforeach; ?>