+++ modules/news/src/admin/Module.php
@@ -5,6 +5,11 @@ namespace luya\news\admin;
 use Yii;
 use luya\admin\components\AdminMenuBuilder;
 
+/**
+ * News Admin Module.
+ * 
+ * @author Basil Suter <basil@nadar.io>
+ */
 class Module extends \luya\admin\base\Module
 {
     public $apis = [
@@ -23,16 +28,25 @@ class Module extends \luya\admin\base\Module
         ],
     ];
     
+    /**
+     * @inheritdoc
+     */
     public function getMenu()
     {
         return (new AdminMenuBuilder($this))
             ->node('news', 'local_library')
                 ->group('news_administrate')
                     ->itemApi('article', 'newsadmin/article/index', 'edit', 'api-news-article')
-                    ->itemApi('cat', 'newsadmin/cat/index', 'bookmark_border', 'api-news-cat')
-                    ->itemApi('tag', 'newsadmin/tag/index', 'label_outline', 'api-news-tag');
+                    ->itemApi('cat', 'newsadmin/cat/index', 'bookmark_border', 'api-news-cat');
     }
 
+    /**
+     * Translat news messages.
+     * 
+     * @param string $message
+     * @param array $params
+     * @return string
+     */
     public static function t($message, array $params = [])
     {
         return Yii::t('newsadmin', $message, $params, Yii::$app->luyaLanguage);