+++ modules/news/src/admin/apis/TagController.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace luya\news\admin\apis;
-
-class TagController extends \luya\admin\ngrest\base\Api
-{
-    public $modelClass = 'luya\news\models\Tag';
-}