+++ modules/news/src/models/Article.php
@@ -2,8 +2,14 @@
 
 namespace luya\news\models;
 
-use luya\news\admin\Module;
 use Yii;
+use yii\helpers\Inflector;
+use luya\helpers\Url;
+use luya\news\admin\Module;
+use luya\admin\aws\TagActiveWindow;
+use luya\admin\ngrest\base\NgRestModel;
+use luya\admin\traits\SoftDeleteTrait;
+use luya\admin\traits\TagsTrait;
 
 /**
  * This is the model class for table "news_article".
@@ -24,9 +30,16 @@ use Yii;
  * @property integer $is_deleted
  * @property integer $is_display_limit
  * @property string $teaser_text
+ * @property string $detailUrl Return the link to the detail url of a news item.
+ * @author Basil Suter <basil@nadar.io>
  */
-class Article extends \luya\admin\ngrest\base\NgRestModel
+class Article extends NgRestModel
 {
+    use SoftDeleteTrait;
+    use TagsTrait;
+    
+    public $i18n = ['title', 'text', 'teaser_text', 'image_list'];
+    
     /**
      * @inheritdoc
      */
@@ -45,14 +58,33 @@ class Article extends \luya\admin\ngrest\base\NgRestModel
         $this->on(self::EVENT_BEFORE_UPDATE, [$this, 'eventBeforeUpdate']);
     }
 
+    public function eventBeforeUpdate()
+    {
+        $this->update_user_id = Yii::$app->adminuser->getId();
+        $this->timestamp_update = time();
+    }
+    
+    public function eventBeforeInsert()
+    {
+        $this->create_user_id = Yii::$app->adminuser->getId();
+        $this->update_user_id = Yii::$app->adminuser->getId();
+        $this->timestamp_update = time();
+        if (empty($this->timestamp_create)) {
+            $this->timestamp_create = time();
+        }
+        if (empty($this->timestamp_display_from)) {
+            $this->timestamp_display_from = time();
+        }
+    }
+    
     /**
      * @inheritdoc
      */
     public function scenarios()
     {
         return [
-           'restcreate' => ['title', 'text', 'cat_id', 'image_id', 'image_list', 'tags', 'timestamp_create', 'timestamp_display_from', 'timestamp_display_until', 'is_display_limit', 'file_list', 'teaser_text'],
-           'restupdate' => ['title', 'text', 'cat_id', 'image_id', 'image_list', 'tags', 'timestamp_create', 'timestamp_display_from', 'timestamp_display_until', 'is_display_limit', 'file_list', 'teaser_text'],
+           'restcreate' => ['title', 'text', 'cat_id', 'image_id', 'image_list', 'timestamp_create', 'timestamp_display_from', 'timestamp_display_until', 'is_display_limit', 'file_list', 'teaser_text'],
+           'restupdate' => ['title', 'text', 'cat_id', 'image_id', 'image_list', 'timestamp_create', 'timestamp_display_from', 'timestamp_display_until', 'is_display_limit', 'file_list', 'teaser_text'],
        ];
     }
 
@@ -69,6 +101,9 @@ class Article extends \luya\admin\ngrest\base\NgRestModel
         ];
     }
 
+    /**
+     * @inheritdoc
+     */
     public function attributeLabels()
     {
         return [
@@ -85,6 +120,9 @@ class Article extends \luya\admin\ngrest\base\NgRestModel
         ];
     }
     
+    /**
+     * @inheritdoc
+     */
     public function ngRestAttributeTypes()
     {
         return [
@@ -102,34 +140,54 @@ class Article extends \luya\admin\ngrest\base\NgRestModel
         ];
     }
 
-    public function eventBeforeUpdate()
+    /**
+     * 
+     * @return string
+     */
+    public function getDetailUrl()
     {
-        $this->update_user_id = Yii::$app->adminuser->getId();
-        $this->timestamp_update = time();
+        return Url::toRoute(['/news/default/detail', 'id' => $this->id, 'title' => Inflector::slug($this->title)]);
     }
 
-    public function eventBeforeInsert()
+    /**
+     * @inheritdoc
+     */
+    public static function ngRestApiEndpoint()
     {
-        $this->create_user_id = Yii::$app->adminuser->getId();
-        $this->update_user_id = Yii::$app->adminuser->getId();
-        $this->timestamp_update = time();
-        if (empty($this->timestamp_create)) {
-            $this->timestamp_create = time();
-        }
-        if (empty($this->timestamp_display_from)) {
-            $this->timestamp_display_from = time();
-        }
+        return 'api-news-article';
     }
-
-    public function getDetailUrl($contextNavItemId = null)
+    
+    /**
+     * @inheritdoc
+     */
+    public function ngRestAttributeGroups()
     {
-        if ($contextNavItemId) {
-            return \luya\cms\helpers\Url::toMenuItem($contextNavItemId, 'news/default/detail', ['id' => $this->id, 'title' => \yii\helpers\Inflector::slug($this->title)]);
-        }
-
-        return \luya\helpers\Url::toManager('news/default/detail', ['id' => $this->id, 'title' => \yii\helpers\Inflector::slug($this->title)]);
+        return [
+            [['timestamp_create', 'timestamp_display_from', 'is_display_limit', 'timestamp_display_until'], 'Time', 'collapsed'],
+            [['image_id', 'image_list', 'file_list'], 'Media'],
+        ];
     }
-
+    
+    /**
+     * @inheritdoc
+     */
+    public function ngRestConfig($config)
+    {
+        $config->aw->load(['class' => TagActiveWindow::class]);
+    
+        $this->ngRestConfigDefine($config, 'list', ['cat_id', 'title', 'timestamp_create', 'image_id']);
+        $this->ngRestConfigDefine($config, ['create', 'update'], ['cat_id', 'title', 'teaser_text', 'text', 'timestamp_create', 'timestamp_display_from', 'is_display_limit', 'timestamp_display_until', 'image_id', 'image_list', 'file_list']);
+    
+        $config->delete = true;
+    
+        return $config;
+    }
+    
+    /**
+     * 
+     * @param string $limit
+     * @return unknown
+     */
     public static function getAvailableNews($limit = false)
     {
         $q = self::find()->where('timestamp_display_from <= :time', ['time' => time()])->orderBy('timestamp_display_from DESC');
@@ -150,45 +208,24 @@ class Article extends \luya\admin\ngrest\base\NgRestModel
         return $articles;
     }
 
-    public function getCategoryName()
+    /**
+     * 
+     * @return \yii\db\ActiveQuery
+     */
+    public function getCat()
     {
-        $catModel = Cat::find()->where(['id' => $this->cat_id])->one();
-
-        return $catModel->title;
-    }
-
-    // ngrest
-
-    public $tags = [];
-
-    public $i18n = ['title', 'text', 'teaser_text', 'image_list'];
-
-    public function extraFields()
-    {
-        return ['tags'];
-    }
-
-    public static function ngRestApiEndpoint()
-    {
-        return 'api-news-article';
+        return $this->hasOne(Cat::class, ['id' => 'cat_id']);
     }
     
-    public function ngRestAttributeGroups()
+    /**
+     * The cat name short getter.
+     *
+     * @return string
+     */
+    public function getCategoryName()
     {
-        return [
-            [['timestamp_create', 'timestamp_display_from', 'is_display_limit', 'timestamp_display_until'], 'Time', 'collapsed'],
-            [['image_id', 'image_list', 'file_list'], 'Media'],
-        ];
+        return $this->cat->title;
     }
 
-    public function ngRestConfig($config)
-    {
-        $this->ngRestConfigDefine($config, 'list', ['title', 'cat_id', 'timestamp_create', 'image_id']);
-
-        $this->ngRestConfigDefine($config, ['create', 'update'], ['cat_id', 'title', 'teaser_text', 'text', 'timestamp_create', 'timestamp_display_from', 'is_display_limit', 'timestamp_display_until', 'image_id', 'image_list', 'file_list']);
-        
-        $config->delete = true;
-
-        return $config;
-    }
+    
 }